# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A production-ready Go backend template built on clean architecture principles.\nFeatures include JWT authentication with multiple algorithms (HMAC, RSA, ECDSA),\nmulti-language validation (English/Indonesian), RabbitMQ message queuing,\nRedis caching with LZ4 compression, and comprehensive error handling.",
        "title": "Ichi-Go API",
        "contact": {
            "name": "API Support",
            "email": "support@ichi-go.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/ichi-go/api",
    "paths": {
...
        "/v1/rbac/roles/{roleId}/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all users that have been assigned a specific role in a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC - User Roles"
                ],
                "summary": "Get users with role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID (uses request context if not provided)",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ichi-go_pkg_utils_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.GetUsersWithRoleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/rbac/users/{userId}/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all roles assigned to a specific user in a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC - User Roles"
                ],
                "summary": "Get user roles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID (uses request context if not provided)",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ichi-go_pkg_utils_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.GetUserRolesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign a role to a user in a specific tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC - User Roles"
                ],
                "summary": "Assign role to user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assign role request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.AssignRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ichi-go_pkg_utils_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/rbac/users/{userId}/roles/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve only active (non-expired) roles for a specific user in a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC - User Roles"
                ],
                "summary": "Get active user roles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID (uses request context if not provided)",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ichi-go_pkg_utils_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.GetUserRolesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/rbac/users/{userId}/roles/{roleSlug}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke a role assignment from a user in a specific tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC - User Roles"
                ],
                "summary": "Revoke role from user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role slug",
                        "name": "roleSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Revoke role request (optional tenant_id and reason)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.RevokeRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ichi-go_pkg_utils_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ichi-go_internal_applications_rbac_dto.AssignRoleRequest": {
            "type": "object",
            "required": [
                "role_slug",
                "tenant_id",
                "user_id"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "reason": {
                    "type": "string",
                    "maxLength": 500
                },
                "role_slug": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "ichi-go_internal_applications_rbac_dto.GetUserRolesResponse": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.UserRoleResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ichi-go_internal_applications_rbac_dto.GetUsersWithRoleResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ichi-go_internal_applications_rbac_dto.UserRoleResponse"
                    }
                }
            }
        },
        "ichi-go_internal_applications_rbac_dto.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "ichi-go_internal_applications_rbac_dto.RevokeRoleRequest": {
            "type": "object",
            "required": [
                "role_slug",
                "tenant_id",
                "user_id"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 500
                },
                "role_slug": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "ichi-go_internal_applications_rbac_dto.UserRoleResponse": {
            "type": "object",
            "properties": {
                "assigned_at": {
                    "type": "string"
                },
                "assigned_by": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "role_id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                },
                "role_slug": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "ichi-go_pkg_utils_response.ErrorDetail": {
            "description": "Detailed error information",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Validation failed"
                }
            }
        },
        "ichi-go_pkg_utils_response.ErrorResponse": {
            "description": "Standard error response wrapper",
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/ichi-go_pkg_utils_response.ErrorDetail"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "ichi-go_pkg_utils_response.SuccessResponse": {
            "description": "Standard success response wrapper",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}